src/requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
Django==2.1
django-cors-headers==2.4.0
django-jquery==3.1.0
django-markdown-deux==1.0.5
django-pagedown==1.0.6
django-tinymce4-lite==1.7.4
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
isort==4.3.4
jsmin==2.2.2
lazy-object-proxy==1.3.1
markdown2==2.3.7
mccabe==0.6.1
Pillow==5.4.1
pipenv==2018.11.26
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pylint==2.2.2
pytz==2018.9
six==1.12.0
uWSGI==2.0.17.1
virtualenv==16.3.0
virtualenv-clone==0.5.1
wrapt==1.11.0
